add read ops sampler
